+++ 02_daemonset.tf
@@ -437,6 +437,7 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
                 mount_path        = volume_mount.value.path
                 mount_propagation = volume_mount.value.propagation
                 name              = volume_mount.key
+                read_only         = volume_mount.value.readOnly
               }
             }
 
@@ -830,6 +831,7 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
                 mount_path        = volume_mount.value.path
                 mount_propagation = volume_mount.value.propagation
                 name              = volume_mount.key
+                read_only         = volume_mount.value.readOnly
               }
             }