+++ 03_statefulset.tf
@@ -465,6 +465,7 @@ resource "kubernetes_stateful_set_v1" "statefulset" {
                 mount_path        = volume_mount.value.path
                 mount_propagation = volume_mount.value.propagation
                 name              = volume_mount.key
+                read_only         = volume_mount.value.readOnly
               }
             }
 
@@ -866,6 +867,7 @@ resource "kubernetes_stateful_set_v1" "statefulset" {
                 mount_path        = volume_mount.value.path
                 mount_propagation = volume_mount.value.propagation
                 name              = volume_mount.key
+                read_only         = volume_mount.value.readOnly
               }
             }