+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - allow configuration of `internalTrafficPolicy` for `clusterIP` services
 - add new optional variable `infraOverrideConfig` that allows overriding defaults supplied through `infrastructureSize`
 - `infrastructureSize` no longer supports additional optional dynamic resource multiplicator introduces in `1.6.0`
+- removed unused variables `operatingSystem` and `architecture`
 
 ## [2.4.0] - 2025-01-01