+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - `infrastructureSize` no longer supports additional optional dynamic resource multiplicator introduces in `1.6.0`
 - removed unused variables `operatingSystem` and `architecture`
 - in repository example template now supports pvc override with `hostPath`
+- add lifecycle.ignore_changes to `pvc` resource
+  - pvc adjustments should be planned and implemented manually
+
 
 ## [2.4.0] - 2025-01-01